Created StickyHeaderScrollView
Moved out from EventPage
