+++ src/components/StickyHeaderScrollView/index.ts
@@ -1,0 +1,3 @@
+import StickyHeaderScrollView from './StickyHeaderScrollView';
+
+export default StickyHeaderScrollView;
