+++ src/components/index.tsx
@@ -2,3 +2,4 @@ export { default as Box } from './Box';
 export { default as Text } from './Text';
 export { default as CircularButton } from './CircularButton';
 export { default as EventBox } from './EventBox';
+export { default as StickyHeaderScrollView } from './StickyHeaderScrollView';
